Fixed sharing
* Setup script adds local recordings folder
* GUI pops when recording retrieval mode is done
* Add timeout check for recording module to see if file is actually created

+++ main_module/dashRecording/DashcamRecorder.py
@@ -208,19 +208,27 @@ class Recorder:
     def wrapping_thread_func(self):
         while not self.terminate_event.is_set():
             while not self.wrapping_queue.empty() and not self.terminate_event.is_set():
-                # check if file is ready
+
                 file_to_wrap_path = self.wrapping_queue.get()
                 if(not self.silent):
                     print("WRAPPING " + file_to_wrap_path)
-                # 'xxx.h264'
-                file_out_path = file_to_wrap_path[:-4] + 'mp4'
-                cmd_str = 'ffmpeg -framerate {0} -i {1} -c:v copy -f mp4 {2}'.format(self.framerate,
-                                                                                     file_to_wrap_path,
-                                                                                     file_out_path)
-                cmd = cmd_str.split()
-                subprocess.call(cmd, stdout=open(os.devnull, 'wb'), stderr=open(os.devnull, 'wb'))
-                os.remove(file_to_wrap_path)
-                self.check_reduce()
+
+                timeout = 5*10
+                # check if file is ready
+                while not os.path.exists(file_to_wrap_path) and timeout > 0:
+                    self.terminate_event.wait(0.1)
+                    timeout = timeout - 1
+
+                if timeout > 0:
+                    # 'xxx.h264'
+                    file_out_path = file_to_wrap_path[:-4] + 'mp4'
+                    cmd_str = 'ffmpeg -framerate {0} -i {1} -c:v copy -f mp4 {2}'.format(self.framerate,
+                                                                                         file_to_wrap_path,
+                                                                                         file_out_path)
+                    cmd = cmd_str.split()
+                    subprocess.call(cmd, stdout=open(os.devnull, 'wb'), stderr=open(os.devnull, 'wb'))
+                    os.remove(file_to_wrap_path)
+                    self.check_reduce()
 
 
             self.check_reduce()
